작업 일시 : 2021.9.14 13:25 작업 내용 : ResourceDictionary.xaml 삭제

+++ ECS.UI/ViewModel/MainWindowViewModel.cs
@@ -748,7 +748,7 @@ namespace ECS.UI.ViewModel
         {
             _engine.Stop();
 
-            if(_ManagedProcess != null)
+            if(_ManagedProcess != null && Process.GetProcessById(_ManagedProcess.Id) != null)
                 _ManagedProcess.Kill();
         }
